requirements.txt
numpy==1.19.5
streamlit==1.0.0
fbprophet==0.5
matplotlib==3.1.3
plotly==5.3.1
DateTime==4.3
pandas==1.0.1
Pillow==7.0.0
